Rails Badassery to keep data in seperate tables
A student model in rails inherits the user model.
It's table is still the users table but in ruby it is a different class.

To get extra fields (without using STI), we created a student_profile model which
'belongs' to the student model.

We then delegate all the getting/setting of the student fields back to the profile.

SCHWEET