Hooked up more navigations.
Ensured that add event is not shown for viewing a company using editable: false
to the partial.
Changed events fetch to only get ones with company id.
Added on destroy event to events index, when an event is destroyed and
removed from the database, the model is destroyed and this bubbles up to the
collection (the events index).