Companies index changes depending on user type
CPP.CurrentUser is now available globally and can be used as a rough guide to
what to display. The back end will hold main authority and reject any bad requests
but for front end A/B views this will be useful.

In this case, I check the user type for the company index view