Abstracted change password into own view
Change password form will be in several different settings pages, so
I have made it easy to create with it's own view.

Added departments router and department settings page. This page
contains a change password form, but this does not currently work as
we don't have department log in yet.