Prettified event signup
Capacity section is now titled 'Sign Up'.

Moved attend/unattend button to Sign Up section and changed the text displayed in the button to a more intuitive 'Sign me up!' and 'I don't want to go'.

Added a bit of text by the attend/unattend button to make it clear to the student what they are doing: 'You are (not) attending this event'

Progress bar now also updates live as students register/unregister.