DepartmentRegistration as discussed
The default value for a department registration is 0 (not requested), meaning that for any departments not selected, value 0 is essentially in the department registration table (created on get/set)

When a company is created, we create a department registration with status 1 (pending) for each department they have selected.

Department admins can see a list of companies with pending status on their dash, with link to edit company.

On edit company screen, department admins can change registration status.

On company dash there is a list of all departments in the system, with their registration status.

Fixes #258