Upgrade backbone-on-rails.
This fixes a problem that occurred when the server responded with a
HTTP 304 Not Modified response. backbone.js would interpret this as a
failure, giving error messages such as "Couldn't fetch student" when
trying to view your own profile.

Later versions of backbone.js fix this problem. I have conservatively
upgraded to the next available version, rather than the current
version, in case of breakage.